app1/requirements.txt
fastapi~=0.117.1
gunicorn~=23.0.0
httpx~=0.28.1
python-dateutil~=2.9.0.post0
python-dotenv~=1.1.1
requests~=2.32.5
pydantic-ai-slim[openai,logfire]==1.18.0